backend/requirements.txt
# ============================================================================
# ShopWhatYouSee - Backend Requirements
# ============================================================================

# Web Framework & API
flask>=2.3.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
werkzeug>=2.3.0

# Deep Learning & Model Loading
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
tensorflow>=2.12.0

# Computer Vision & Image Processing
opencv-python-headless>=4.8.0
Pillow>=9.5.0

# Data Processing & ML Utilities
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.10.0

# Search & Indexing
faiss-cpu>=1.7.4

# Database (PostgreSQL)
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Utilities
requests>=2.31.0
tqdm>=4.65.0
pydantic>=2.0.0
pyyaml>=6.0

# Evaluation & Visualization (optional, for notebooks)
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0

# Logging & Monitoring (optional)
python-logging-loki>=0.3.2
